переписал роут, перенес в разные папки разные компоненты и функции, остальное в комментариях в файлах
